rename: ThumbnailBox 공통컴포넌트로 이동
