+++ src/src/components/common/images/ThumbnailBox.jsx
